+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 0.4
 ---------------------------
 - Added ``AverageModels`` class to enable averaging across multiple fits.
+- Added ``polyclonal_collection.fit_models`` to fit multiple models using multiprocessing.
 
 0.3
 ---------------------------